+++ env/webpack_setup.sh
@@ -1,0 +1,2 @@
+echo "Installing npm packages"
+yarn