+++ zapisy/apps/offer/preferences/views.py
@@ -7,8 +7,8 @@ from django.views.decorators.http import require_POST
 from libs.ajax_messages import AjaxFailureMessage
 from apps.offer.preferences.forms import PreferenceFormset, PreferenceForm
 from apps.offer.preferences.models import Preference
-from apps.users.decorators import employee_required
 from apps.enrollment.courses.models.semester import Semester
+from apps.users.decorators import employee_required
 
 logger = logging.getLogger()