+++ zapisy/apps/users/exceptions.py
@@ -3,16 +3,6 @@ class NonUserException(Exception):
     pass
 
 
-class NonEmployeeException(Exception):
-    """Thrown when a user is not an employee"""
-    pass
-
-
-class NonStudentException(Exception):
-    """Thrown when a user is not a student"""
-    pass
-
-
 class InvalidUserException(Exception):
     """
     Thrown when a user is considered to be invalid