+++ zapisy/asset-defs.ts
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
 	rawfiles: [
 		{ from: "legacy/css/", to: "css/" },
 		{ from: "legacy/help/", to: "help/" },
@@ -6,6 +6,11 @@ module.exports = {
 		{ from: "legacy/js/", to: "js/" },
 		{ from: "legacy/vendor/", to: "vendor/" },
 		{ from: "legacy/favicon.ico", to: "favicon.ico" },
-		{ from: "legacy/feed-icon.png", to: "feed-icon.png" }
+		{ from: "legacy/feed-icon.png", to: "feed-icon.png" },
 	],
+	bundles: {
+		"main": [
+			"common/libs.ts",
+		],
+	},
 };