+++ zapisy/assets/common/utils.ts
@@ -1,9 +1,9 @@
 // Common utility functions shared between apps
-
 import * as $ from "jquery";
+import { UnconstrainedFunction } from "./types";
 
-function scrollUpToElementIfWindowBelow(selector: string) {
-	const SCROL_TIME = 400;
+const SCROLL_TIME = 400;
+export function scrollUpToElementIfWindowBelow(selector: string): void {
 	const currentWindowTop = $(window).scrollTop();
 	if (!currentWindowTop) {
 		return;
@@ -14,8 +14,20 @@ function scrollUpToElementIfWindowBelow(selector: string) {
 	}
 	const elemTop = elem.offset()!.top;
 	if (currentWindowTop > elemTop) {
-		$("html, body")!.animate({ scrollTop: elemTop }, SCROL_TIME);
+		$("html, body").animate({ scrollTop: elemTop }, SCROLL_TIME);
 	}
 }
 
-export { scrollUpToElementIfWindowBelow };
+export function whenDomLoaded(fn: UnconstrainedFunction): void {
+	if (document.readyState !== "loading") {
+		fn();
+	} else {
+		window.addEventListener("DOMContentLoaded", fn);
+	}
+}
+
+export function awaitSleep(ms: number): Promise<void> {
+	return new Promise((resolve, _) => {
+		window.setTimeout(resolve, ms);
+	});
+}